Move "smells like mastodon" url check to click handler
